Fixed Memory Leak Associated with Pools
The library should not use the handleDisconnect function when working with pools because it will cause the following error message to arise over time:

"warning: possible EventEmitter memory leak detected. 11 listeners added."

The pool now handles the opening and closing of connections, so we don't really need this for connection pools.